move rewards.ts inside sdk/util
