+++ subgraphs/stargate/protocols/stargate/config/deployments/stargate-polygon/configurations.ts
@@ -1,9 +1,9 @@
 import { Configurations } from "../../../../../configurations/configurations/interface";
 import {
-  Network,
   PROTOCOL_NAME,
   PROTOCOL_SLUG,
 } from "../../../../../src/common/constants";
+import { Network } from "../../../../../src/sdk/util/constants";
 
 export class StargatePolygonConfigurations implements Configurations {
   getNetwork(): string {
