+++ subgraphs/stargate/src/common/constants.ts
@@ -1,48 +1,13 @@
 import { Address, BigInt, TypedMap } from "@graphprotocol/graph-ts";
 
-////////////////////
-///// Versions /////
-////////////////////
-
-export const PROTOCOL_NAME = "Stargate Finance";
-export const PROTOCOL_SLUG = "stargate";
-
-////////////////////////
-///// Schema Enums /////
-////////////////////////
-
-// The network names corresponding to the Network enum in the schema.
-// They also correspond to the ones in `dataSource.network()` after converting to lower case.
-// See below for a complete list:
-// https://thegraph.com/docs/en/hosted-service/what-is-hosted-service/#supported-networks-on-the-hosted-service
-export namespace Network {
-  export const ARBITRUM_ONE = "ARBITRUM_ONE";
-  export const ARWEAVE_MAINNET = "ARWEAVE_MAINNET";
-  export const AURORA = "AURORA";
-  export const AVALANCHE = "AVALANCHE";
-  export const BOBA = "BOBA";
-  export const BSC = "BSC"; // aka BNB Chain
-  export const CELO = "CELO";
-  export const COSMOS = "COSMOS";
-  export const CRONOS = "CRONOS";
-  export const MAINNET = "MAINNET"; // Ethereum mainnet
-  export const FANTOM = "FANTOM";
-  export const FUSE = "FUSE";
-  export const HARMONY = "HARMONY";
-  export const JUNO = "JUNO";
-  export const MOONBEAM = "MOONBEAM";
-  export const MOONRIVER = "MOONRIVER";
-  export const NEAR_MAINNET = "NEAR_MAINNET";
-  export const OPTIMISM = "OPTIMISM";
-  export const OSMOSIS = "OSMOSIS";
-  export const MATIC = "MATIC"; // aka Polygon
-  export const XDAI = "XDAI"; // aka Gnosis Chain
-  export const UNKNOWN_NETWORK = "UNKNOWN_NETWORK";
-}
+import { Network } from "../sdk/util/constants";
 
 /////////////////////////////
 ///// Protocol Specific /////
 /////////////////////////////
+
+export const PROTOCOL_NAME = "Stargate Finance";
+export const PROTOCOL_SLUG = "stargate";
 
 export const crossPoolTokens = new TypedMap<
   string,
